requirements.txt
Django>=5.0,<6.0
django-ninja>=1.3,<2.0
djangorestframework
djangorestframework-simplejwt
celery
redis
email-validator
django-ninja-jwt
djangorestframework-jwt
pydantic[email]